Add some comments, refactor the remaining components out of the App.js file, make the game information card render description as interpreted html.
